+++ src/stacks/CharacterDetails/screens/CharacterDetails/CharacterDetails.screen.tsx
@@ -3,7 +3,7 @@ import {Text, View, Image } from 'react-native';
 import { atom, useAtomValue, useSetAtom } from 'jotai';
 import { CharacterInfo, selectedCharacterAtom } from '../../../CharacterList/screens/CharacterList/CharacterList.screen';
 import LikeButton from '../../../../components/LikeButton/LikeButton';
-import * as api from '../../../../fetch';
+import * as api from '../../../../api';
 import {styles} from './CharacterDetails.styled';
 
 const characterInfoAtom = atom<CharacterInfo>({} as CharacterInfo)
